requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
pytest==7.1.2
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-parallel==0.1.1
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
selenium==4.1.5
requests==2.27.1
urllib3==1.24
numpy==1.22.3
PyYAML==6.0
loguru==0.6.0
opencv-python==4.5.5.64

Faker==13.11.1
xlrd==2.0.1